+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix #6247: Support for proxy authentication from proxy URL user info
 * Fix #6342: UnmatchedFieldTypeModule prevents certain jackson features from working
 * Fix #6354: Prevent deadlock in okhttp AsyncBody.cancel
+* Fix #6366: Allow Retry-After header to be considered in retries
 
 ### 6.13.3 (2024-08-13)